pyproject.toml
[project]
name = "portfolio_backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Atrayan Mukherjee",email = "atrayanmukherjee0248@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "django (>=5.1.5,<6.0.0)",
    "django-rest-framework (>=0.1.0,<0.2.0)",
    "django-cors-headers (>=4.6.0,<5.0.0)",
    "django-filter (>=24.3,<25.0)",
    "django-split-settings (>=1.3.2,<2.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "pymongo[gssapi,snappy,srv,tls] (>=4.11,<5.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pillow (>=11.1.0,<12.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.1.0"
types-pyyaml = "^6.0.12.20241230"
colorlog = "^6.9.0"

[tool.isort]
multi_line_output = 5
line_length =119

[tool.yapf]
based_on_style = "google"
align_closing_bracket_with_visual_indent = true
coalesce_brackets = true
column_list = 119
dedent_closing_brackets = true
